requirements.txt
# requirements.txt
pandas
numpy
fastapi
scikit-learn
joblib
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
